Add current date to todo list
This commit adds a date to a Todo list. This is done using dateHelpers
so that the code involved in the Todo app is only related to the
TodoApp.

The `dateHelper` is crude but serves as an example to how helpers might
be used to abstract code away. No tests exist yet but they would be
added if more time was available.