Add ability to add new todo items to the list
This commit includes adding the controlled form input to allow new todo
items to be added to the todo state array. The ID for each todo is not
ideal and a new system would be preferably but it's good enough for this
initial set up.

The feature spec needs making more resilient but that can come in later
commits.