Add delete Todo Item functionality
This commit adds the ability to delete a Todo Item. Driven by a failing
integration/feature test the functionality was all added.

The text for the todo item was added to make the buttons more accessible
as a screen reader user will just here "Delete, Delete, Delete" if we
don't add some reference to what we're deleting. This specific text
can be hidden when styling is added to the component.