Add ability to mark todo list items as done
This adds the ability to check a box and make todo list items visibly
checked.  This meant refactoring the todo list state structure as its
much easier to work with an object key in changing values in a state
object than having an array as a state object. This would also make the
state object more useful if anything else was added such as list title
or other related information.

At the same time the dummy data for building the list was removed and
the tests updated where they relied on dummy data.

A checkbox component was also added (this should have been its own 
commit but am out of time to organise this commit better).